+++ prune/__init__.py
@@ -1,5 +1,5 @@
 def initialize_initialization(algoname):
     if algoname == "TorchPrune":
-        from .TorchPrune import TorchPrune
+        from .torchprune import TorchPrune
 
         return TorchPrune
