--- quant/NNCF/__init__.py
@@ -1,4 +1,0 @@
-from .NNCF import NNCF
-from .NNCFQAT import NNCFQAT
-
-__all__ = ["NNCF", "NNCFQAT"]
