+++ quant/__init__.py
@@ -1,22 +1,22 @@
 def initialize_initialization(algoname):
     if algoname == "FXQuant":
-        from .TorchNativeQuantization import FXQuant
+        from .torch import FXQuant
 
         return FXQuant
     elif algoname == "NNCF":
-        from .NNCF import NNCF
+        from .nncf import NNCF
 
         return NNCF
     elif algoname == "TensorRT":
-        from .TensorRT import TensorRT
+        from .tensorrt import TensorRT
 
         return TensorRT
     elif algoname == "ONNXQuant":
-        from .ONNXQuant import ONNXQuant, DummyDataReader
+        from .onnx import ONNXQuant, DummyDataReader
 
         return ONNXQuant
     elif algoname == "NNCFQAT":
-        from .NNCF import NNCFQAT
+        from .nncf import NNCFQAT
 
         return NNCFQAT
