+++ quant/nncf/__init__.py
@@ -1,0 +1,4 @@
+from .nncf import NNCF
+from .nncfqat import NNCFQAT
+
+__all__ = ["nncf", "nncfqat"]
