+++ scripts/prune/TorchPrune.yaml
@@ -1,7 +1,6 @@
 ALGORITHM: TorchPrune
 ALGO_TYPE: prune
 BATCH_SIZE: 512
-DATA_URL: ''
 DATASET_NAME: CIFAR10
 MODEL: resnet50
 NUM_CLASSES: 10
@@ -9,7 +8,6 @@ CACHE_PATH: 'ABC/.cache'
 JOB_ID: 1
 JOB_PATH: 'ABC/jobs/1'
 DATA_PATH: 'ABC/datasets/1'
-PIN_MEM: false
 PRETRAINED: false
 PLATFORM: torchvision
 TASK: image_classification
@@ -26,7 +24,6 @@ LEARNING_RATE: 0.001
 FINETUNE_EPOCHS: 1
 VALIDATE: True
 VALIDATION_INTERVAL: 1
-USER_FOLDER: 'ABC'
 LOGGING_PATH: 'ABC/logs' 
 CUSTOM_MODEL_PATH: 'ABC/models'
 MODEL_PATH: 'ABC/jobs/1'
