+++ scripts/prune/TorchPruneVIT.yaml
@@ -1,7 +1,6 @@
 ALGORITHM: TorchPrune
 ALGO_TYPE: prune
 BATCH_SIZE: 32
-DATA_URL: ''
 DATASET_NAME: CIFAR10 
 MODEL: vit_small_patch16_224
 NUM_CLASSES: 10
@@ -11,7 +10,6 @@ JOB_PATH: 'ABC/jobs/1'
 DATA_PATH: 'ABC/datasets/1'
 LOGGING_PATH: 'ABC/logs' 
 MODEL_PATH: 'ABC/models'
-PIN_MEM: false
 PRETRAINED: false
 PLATFORM: timm
 TASK: image_classification
@@ -28,7 +26,7 @@ LEARNING_RATE: 0.001
 FINETUNE_EPOCHS: 0
 VALIDATE: True
 VALIDATION_INTERVAL: 1
-USER_FOLDER: 'ABC'
+
 prune:
   TorchPrune:
       GROUP_IMPORTANCE: GroupTaylorImportance
