+++ scripts/quant/FXQuantQAT.yaml
@@ -1,19 +1,16 @@
 ALGORITHM: FXQuant
 ALGO_TYPE: quant
 BATCH_SIZE: 32
-DATA_URL: ''
-DATASET_NAME: CIFAR10 # incase of autogptq ["wikitext2", "c4", "c4-new", "ptb", "ptb-new"]
+DATASET_NAME: CIFAR10 
 MODEL: resnet50
 NUM_CLASSES: 10
 CACHE_PATH: 'ABC/.cache'
 JOB_ID: 1
 JOB_PATH: 'ABC/jobs/1'
 DATA_PATH: 'ABC/datasets/1'
-PIN_MEM: False
 PRETRAINED: False
 PLATFORM: torchvision
 TASK: image_classification
-WORKERS: 0
 insize: 32
 wandb: False 
 VERSION: original
@@ -26,7 +23,6 @@ LEARNING_RATE: 0.001
 FINETUNE_EPOCHS: 1
 VALIDATE: True
 VALIDATION_INTERVAL: 1
-USER_FOLDER: 'ABC'
 LOGGING_PATH: 'ABC/logs' 
 CUSTOM_MODEL_PATH: 'ABC/models'
 MODEL_PATH: 'ABC/jobs/1'
@@ -35,7 +31,6 @@ quant:
           choice: 'qat' # weight/static/fusion/qat
           EPOCHS: 1
           OPTIMIZER: 'adam'
-          LOSS_FN: 'ce'
-          QUANTIZED_MODEL_PATH: 'quantized_models/fxquant_model.pt'
+          LOSS_FN: 'ce' # ce = cross entropy
