+++ scripts/quant/NNCFPTQ.yaml
@@ -1,19 +1,16 @@
 ALGORITHM: NNCF
 ALGO_TYPE: quant
 BATCH_SIZE: 1
-DATA_URL: ''
-DATASET_NAME: CIFAR10 # incase of autogptq ["wikitext2", "c4", "c4-new", "ptb", "ptb-new"]
+DATASET_NAME: CIFAR10
 MODEL: resnet50
 NUM_CLASSES: 10
 CACHE_PATH: 'ABC/.cache'
 JOB_ID: 1
 JOB_PATH: 'ABC/jobs/1'
 DATA_PATH: 'ABC/datasets/1'
-PIN_MEM: false
 PRETRAINED: false
 PLATFORM: torchvision
 TASK: image_classification
-WORKERS: 0
 insize: 32
 wandb: False 
 VERSION: original
@@ -26,11 +23,9 @@ LEARNING_RATE: 0.001
 FINETUNE_EPOCHS: 1
 VALIDATE: True
 VALIDATION_INTERVAL: 1
-USER_FOLDER: 'ABC'
 LOGGING_PATH: 'ABC/logs' 
 CUSTOM_MODEL_PATH: 'ABC/models'
 MODEL_PATH: 'ABC/jobs/1'
-
 quant:
     NNCF:
         TRANSFORMER: True
