+++ scripts/quant/ONNXPTQ.yaml
@@ -1,8 +1,7 @@
 ALGORITHM: ONNXQuant
 ALGO_TYPE: quant
 BATCH_SIZE: 32
-DATA_URL: ''
-DATASET_NAME: CIFAR10 # incase of autogptq ["wikitext2", "c4", "c4-new", "ptb", "ptb-new"]
+DATASET_NAME: CIFAR10
 MODEL: vit_small_patch16_224
 NUM_CLASSES: 10
 CACHE_PATH: 'ABC/.cache'
@@ -12,7 +11,6 @@ DATA_PATH: 'ABC/datasets/1'
 LOGGING_PATH: 'ABC/logs' 
 CUSTOM_MODEL_PATH: 'ABC/models'
 MODEL_PATH: 'ABC/jobs/1'
-PIN_MEM: false
 PRETRAINED: false
 PLATFORM: timm
 TASK: image_classification
@@ -29,11 +27,9 @@ LEARNING_RATE: 0.001
 FINETUNE_EPOCHS: 0
 VALIDATE: True
 VALIDATION_INTERVAL: 1
-USER_FOLDER: 'ABC'
+
 quant:
     ONNXQuant:
-      onnx_file_path : "onnx_models/resnet50_trial.onnx"
-      onnx_quantized_file_path: "onnx_quantized_models/resnet50_trial.onnx"
       quant_format: "QuantFormat.QDQ"
       per_channel: False 
       activation_type: "QuantType.QInt8"
